requirements.txt
fastapi                    
uvicorn[standard]         
pydantic
pydantic[email]               
pydantic-settings     
python-jose[cryptography]
python-multipart


sqlalchemy[asyncio]        
asyncpg                    
alembic                  
httpx
requests
python-jose[cryptography]   
passlib[bcrypt]             


azure-storage-blob        
azure-ai-contentsafety      

tzdata
reportlab
python-dotenv               
httpx                     
Pillow                 
redis                       
PyPDF2
black

aiofiles